pyproject.toml
[project]
name = "nonebot-plugin-orangejuice"
version = "0.3.0"
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0-beta.1",
    "pymysql>=1.0.0"
]
requires-python = ">=3.8"
authors = [{ name = "Polaris_Light", email = "995905922@qq.com" }]
description = "Sorabot Implemention"
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["orangejuice"]
classifiers = []

[project.urls]
Homepage = "https://github.com/FDCraft/nonebot-plugin-orangejuice"

[tool.poetry.dependencies]
nonebot-adapter-onebot = ">=2.0.0-beta.1"
nonebot2 = ">=2.0.0"
python = "^3.8"
pymysql = ">=1.0.0"
